fixes for hackrf devices not being able to run in multiple nodes on the same host, and some related refactoring of the capture code to make it more modular and easier to maintain. Also added a new config option for CFAR mode in the detection algorithm, and updated the API server to initialize the stashes with the config.

+++ src/capture/hackrf/HackRf.h
@@ -36,7 +36,7 @@ private:
   /// @brief Check status of HackRF API returns.
   /// @param status Return code of API call.
   /// @param message Message if API call error.
-  void check_status(uint8_t status, std::string message);
+  void check_status(int status, std::string message);
 
 protected:
   /// @brief Array of pointers to HackRF devices.
